workerDiego/requirements.txt
accelerate==1.7.0
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
bitsandbytes==0.46.0
boto3==1.38.26
botocore==1.38.26
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
cut-cross-entropy==25.1.1
datasets==3.6.0
diffusers==0.33.1
dill==0.3.8
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.16
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.2
hf_transfer==0.1.9
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.3
idna==3.10
importlib_metadata==8.7.0
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
orjson==3.10.18
packaging==25.0
pandas==2.2.3
paramiko==3.5.1
peft==0.15.2
pfzy==0.3.4
pillow==11.2.1
platformdirs==4.3.8
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycares==4.8.0
pycparser==2.22
pydantic==2.11.5
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
runpod==1.7.10
s3transfer==0.13.0
safetensors==0.5.3
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.6
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
tqdm-loggable==0.2
transformers==4.51.3
triton==3.3.0
trl==0.15.2
typeguard==4.4.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tyro==0.9.22
tzdata==2025.2
ujson==5.10.0
unsloth==2025.5.7
unsloth_zoo==2025.5.11
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
wandb==0.19.11
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
xformers==0.0.30
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
